Functie die data op postcode niveau omzet naar data op wijk- of gemeenteniveau

Functie die data op postcode niveau omzet naar data op wijk- of gemeenteniveau

Usage

vertaal_postcode_naar_regio_op_peiljaar(
  df,
  oorspronkelijk_jaar,
  peiljaar,
  type_kolommen = "aantal",
  regionaalniveau = "wijk"
)

Arguments

df

data.frame op postcode niveau dat omgezet moet worden.

oorspronkelijk_jaar

het jaar waaruit de data in df afkomstig is.

peiljaar

jaar waarnaartoe de features moeten worden omgezet

type_kolommen

welk type is de kolom, 'aantal' of 'aandeel'?

regionaalniveau

op welk regionaalniveau moet de omzetting plaats vinden? Keuze uit 'wijk' en 'gemeente'.

Value

Omgezet data-frame

Examples

library(grenswijzigen)
library(cbsodataR)
library(dplyr)

# laad de bevolkingsgrootte per 4-cijferige postcode
df.postcode <- cbs_get_data(
  id = "83502NED",
  Geslacht = "T001038",
  Leeftijd = "10000",
  Postcode = has_substring("PC"),
  Perioden = "2020JJ00",
  select = c("Geslacht", "Leeftijd", "Postcode", "Perioden", "Bevolking_1")
) %>% mutate(
  Postcode = gsub("\\D", "", Postcode)
) %>% select("Postcode", "Bevolking_1")

# zet om naar wijk
df.vertaald = vertaal_postcode_naar_regio_op_peiljaar(
  df.postcode,
  oorspronkelijk_jaar = 2020,
  peiljaar = 2021,
  type_kolommen = "aantal",
  regionaalniveau = "wijk"
)
#> [1] "Aantal rijen omgezette data-frame: 3243"

print(head(df.vertaald))
#>   gwb_code Bevolking_1 jaar
#> 1   001400   23400.192 2020
#> 2   001401   19864.358 2020
#> 3   001402   14374.962 2020
#> 4   001403   18545.000 2020
#> 5   001404   11841.837 2020
#> 6   001405    3996.121 2020
# gwb_code Bevolking_1 jaar
# 1     1400   23400.192 2020
# 2     1401   19864.358 2020
# 3     1402   14374.962 2020
# 4     1403   18545.000 2020
# 5     1404   11841.837 2020
# 6     1405    3996.121 2020

# laad de bevolkingsgrootte uit de KWB-data
df.wijk <- cbs_get_data(
  id="85039NED",
  WijkenEnBuurten = has_substring("WK"),
  select = c("WijkenEnBuurten", "AantalInwoners_5")
)

print(head(df.wijk %>% arrange(WijkenEnBuurten)))
#> # A tibble: 6 × 2
#>   WijkenEnBuurten AantalInwoners_5
#>   <chr>                      <int>
#> 1 "WK001400  "               22735
#> 2 "WK001401  "               19695
#> 3 "WK001402  "               14055
#> 4 "WK001403  "               18405
#> 5 "WK001404  "               12355
#> 6 "WK001405  "                3290
# # A tibble: 6 x 2
#   WijkenEnBuurten AantalInwoners_5
#   <chr>                      <int>
# 1 "WK001400  "               22730
# 2 "WK001401  "               19695
# 3 "WK001402  "               14055
# 4 "WK001403  "               18410
# 5 "WK001404  "               12355
# 6 "WK001405  "                3290